packages/pdw/pdw-0.5.tar.gz/pdw-0.5/pdw/tests/getdata/test_openlibrary.py
import pkg_resources

from pdw.getdata.openlibrary import *
import pdw.getdata.openlibrary as openlibrary

import pdw.model as model

class TestLoader:
    @classmethod
    def teardown_class(self):
        model.repo.rebuild_db()

    def test_load_authors_and_editions(self):
        loader = openlibrary.Loader()
        # not strictly json -- each line is a json string
        fo = pkg_resources.resource_stream('pdw', 'tests/data/openlibrary.authors.json')
        startnum = model.Person.query.count()
        persons = [ p for p in loader.load_to_db(fo, objtype='person') ]
        model.Session.commit()
        endnum = model.Person.query.count()
        assert endnum == startnum + 4, endnum
        assert persons[-1].name == u'Tolstoy, Leo'
        model.Session.remove()
    
        # don't both with separate load as need author to ensure matching
        fo2 = pkg_resources.resource_stream('pdw', 'tests/data/openlibrary.editions.json')
        items = [ it for it in loader.load_to_db(fo2, objtype='item') ]
        model.Session.commit()
        wandp = items[0]
        assert len(items) == 1
        # in name was Leo Tolstoy but should be normalized ...
        assert wandp.title.startswith(u'War and Peace'), wandp

        tolstoyid = loader.srcid('/a/OL2624954A')
        tolstoy = model.Person.query.filter_by(srcid=tolstoyid).first()
        assert tolstoy.items[0] == items[0]


class TestOLQuery:
    external = True
    __test__ = False
    olq = OLQuery()

    def test_search_author(self):
        return
        # fails now for unknown reason ...
        jkid = self.olq.search_author('J.K. Rowling')[0]
        assert jkid == '/a/OL2703514A', jkid
        jkobj = self.olq.get_object(jkid)
        assert jkobj['name'] == 'J.K. Rowling'

        # fails
        # jkid = self.olq.search_author('Rowling, J.K.')
        # assert jkid == '/a/OL2703514A'

    def test_search_author2(self):
        id = self.olq.search_author('Felix Klein')[0]
        # not the other felix klein ...
        assert id == '/a/OL6029810A', id
        obj = self.olq.get_object(id)
        assert obj['name'] == 'Felix Klein', obj

    def test_get_object(self):
        # Felix Klein
        id = '/a/OL668475A'
        out = self.olq.get_object(id)
        assert out['birth_date'] == u'1849', out

    def test_search_book(self):
        out = self.olq.search_book('A Christmas Carol')
        assert out[0] == '/b/OL1036223M', out
        # author is mis-stated as being Samantha Carol Smith not Dickens!
        # http://openlibrary.org/a/OL556641A
        # obj = self.olq.get_object(out[0])
        # assert obj['authors'][0]['key'] == '/a/OL2764072A', obj

    def test_object_search_book(self):
        out = self.olq.object_search_book(title='A Christmas Carol')
        assert out[0] == '/b/OL21138583M', out
        obj = self.olq.get_object(out[0])
        assert obj['authors'][0]['key'] == '/a/OL2764072A', obj

    def test_object_search_by_title_with_article(self):
        out = self.olq.search_book(title='Secret Garden')
        assert out, out
        out = self.olq.search_book(title='Secret Garden,The')
        assert out, out

    def test_person_is_pd(self):
        dickens =  self.olq.get_object(u'/a/OL24638A')
        assert self.olq.person_is_pd(dickens), dickens

        wallace = self.olq.get_object( u'/a/OL448939A')
        print self.olq.person_is_pd(wallace)
        assert not self.olq.person_is_pd(wallace), wallace
        
    def test_by_work(self):
        class _Person:
            pass
        class _Work:
            pass

        p1 = _Person()
        p1.name = u'Dickens, Charles'
        w = _Work()
        w.title = u'A Christmas Carol'
        w.persons = [ p1 ]
        is_pd, work, details = self.olq.by_work(w)
        # tuple (is_pd, work_id)
        assert is_pd, work

        p = _Person()
        p.name = u'David Foster Wallace'
        w = _Work()
        w.title = u'Infinite Jest'
        w.persons = [ p ]
        is_pd, work, details = self.olq.by_work(w)
        # tuple (is_pd, work_id)
        assert not is_pd, work

